made TeamSettings mobile friendly

+++ resources/assets/js/mixins/Requests.js
@@ -1,10 +1,23 @@
 export default
 {
+	data()
+	{
+		return {
+			outstanding: {
+				gets: [],
+				posts: [],
+				puts: [],
+				deletes: [],
+			},
+		}
+	},
+
+
 	methods:
 	{
 		// send a GET request with the given parameters
 		// execute the given event string when finished
-		get(url, successEvent = null, data = [], failEvent = null, options = {})
+		get(url, successEvent = null, data = [], failEvent = null, options = {}, allowMultiple = false)
 		{
 			var self = this;
 
@@ -12,9 +25,18 @@ export default
 				options.headers = {'X-CSRF-TOKEN' : $('#_token').attr('value')};
 			} 
 
+			if (! allowMultiple && this.outstanding.gets.indexOf(url) !== -1) {
+				// there is already a GET request to this url outstanding
+				return;
+			}
+
+			this.outstanding.gets.push(url);
+
 			this.$http.get(url, data, options)
 				.then(function(response)
 				{
+					this.requestFinished('gets', url);
+
 					if (response.data.ok === false) {
 						throw response.data.error
 					}
@@ -30,6 +52,8 @@ export default
 				})
 				.catch(function(response)
 				{
+					this.requestFinished('gets', url);
+
 					if (failEvent) {
 						self.$broadcast(failEvent, response);
 					}
@@ -42,7 +66,7 @@ export default
 
 		// send a POST request with the given parameters
 		// execute the given event string when finished
-		post(url, successEvent = null, data = [], failEvent = null, options = {})
+		post(url, successEvent = null, data = [], failEvent = null, options = {}, allowMultiple = false)
 		{
 			var self = this;
 
@@ -50,9 +74,18 @@ export default
 				options.headers = {'X-CSRF-TOKEN' : $('#_token').attr('value')};
 			} 
 
+			if (! allowMultiple && this.outstanding.posts.indexOf(url) !== -1) {
+				// there is already a GET request to this url outstanding
+				return;
+			}
+
+			this.outstanding.posts.push(url);
+
 			this.$http.post(url, data, options)
 				.then(function(response)
 				{
+					this.requestFinished('posts', url);
+
 					if (response.data.ok === false) {
 						throw response.data.error
 					}
@@ -68,6 +101,8 @@ export default
 				})
 				.catch(function(response)
 				{
+					this.requestFinished('posts', url);
+
 					if (failEvent) {
 						self.$broadcast(failEvent, response);
 					}
@@ -80,7 +115,7 @@ export default
 
 		// send a PUT request with the given parameters
 		// execute the given event string when finished
-		put(url, successEvent = null, data = [], failEvent = null, options = {})
+		put(url, successEvent = null, data = [], failEvent = null, options = {}, allowMultiple = false)
 		{
 			var self = this;
 
@@ -88,9 +123,18 @@ export default
 				options.headers = {'X-CSRF-TOKEN' : $('#_token').attr('value')};
 			} 
 
+			if (! allowMultiple && this.outstanding.puts.indexOf(url) !== -1) {
+				// there is already a GET request to this url outstanding
+				return;
+			}
+
+			this.outstanding.puts.push(url);
+
 			this.$http.put(url, data, options)
 				.then(function(response)
 				{
+					this.requestFinished('puts', url);
+
 					if (response.data.ok === false) {
 						throw response.data.error
 					}
@@ -106,6 +150,8 @@ export default
 				})
 				.catch(function(response)
 				{
+					this.requestFinished('puts', url);
+
 					if (failEvent) {
 						self.$broadcast(failEvent, response);
 					}
@@ -118,7 +164,7 @@ export default
 
 		// send a DELETE request with the given parameters
 		// execute the given event string when finished
-		delete(url, successEvent = null, data = [], failEvent = null, options = {})
+		delete(url, successEvent = null, data = [], failEvent = null, options = {}, allowMultiple = false)
 		{
 			var self = this;
 
@@ -126,9 +172,18 @@ export default
 				options.headers = {'X-CSRF-TOKEN' : $('#_token').attr('value')};
 			} 
 
+			if (! allowMultiple && this.outstanding.deletes.indexOf(url) !== -1) {
+				// there is already a GET request to this url outstanding
+				return;
+			}
+
+			this.outstanding.deletes.push(url);
+
 			this.$http.delete(url, data, options)
 				.then(function(response)
 				{
+					this.requestFinished('deletes', url);
+
 					if (response.data.ok === false) {
 						throw response.data.error
 					}
@@ -144,6 +199,8 @@ export default
 				})
 				.catch(function(response)
 				{
+					this.requestFinished('deletes', url);
+
 					if (failEvent) {
 						self.$broadcast(failEvent, response);
 					}
@@ -151,6 +208,20 @@ export default
 						self.errorMsg(response);
 					}
 				});
+		},
+
+		/**
+		 * A request has returned, remove this url from the list of outstanding requests
+		 * Each verb has their own array
+		 */
+		requestFinished(verb, url)
+		{
+			let outstandingRequests = this.$get(`outstanding.${verb}`);
+			let index = outstandingRequests.indexOf(url);
+			while (index !== -1) {
+				outstandingRequests.splice(index, 1);
+				index = outstandingRequests.indexOf(url);
+			}
 		},
 	},
 }
